% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{hf_clear_model_cache}
\alias{hf_clear_model_cache}
\title{Clear Cached Model Files}
\usage{
hf_clear_model_cache(
  model_id,
  cache_dir = "~/.cache/huggingface/transformers/",
  warn = TRUE
)
}
\arguments{
\item{model_id}{Model whose files you'd like to remove. Note: partial matches will work.}

\item{cache_dir}{Directory where model files are cached. Defaults to '~/.cache/huggingface/transformers/'.}

\item{warn}{Whether to prompt user with a warning message or simply remove the cached files. Defaults to TRUE.}
}
\description{
Clears cached model files
}
