% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ez.R
\name{hf_ez_conversational_api_inference}
\alias{hf_ez_conversational_api_inference}
\title{Conversational API Inference}
\usage{
hf_ez_conversational_api_inference(
  text,
  generated_responses = NULL,
  past_user_inputs = NULL,
  min_length = NULL,
  max_length = NULL,
  top_k = NULL,
  top_p = NULL,
  temperature = 1,
  max_time = NULL,
  tidy = TRUE,
  use_gpu = FALSE,
  use_cache = FALSE,
  wait_for_model = FALSE,
  use_auth_token = NULL,
  stop_on_error = FALSE,
  ...
)
}
\arguments{
\item{text}{The last input from the user in the conversation.}

\item{generated_responses}{A list of strings corresponding to the earlier replies from the model.}

\item{past_user_inputs}{A list of strings corresponding to the earlier replies from the user. Should be of the same length of generated_responses.}

\item{min_length}{(Default: None). Integer to define the minimum length in tokens of the output summary.}

\item{max_length}{(Default: None). Integer to define the maximum length in tokens of the output summary.}

\item{top_k}{(Default: None). Integer to define the top tokens considered within the sample operation to create new text.}

\item{top_p}{(Default: None). Float to define the tokens that are within the sample operation of text generation. Add tokens in the sample for more probable to least probable until the sum of the probabilities is greater than top_p.}

\item{temperature}{(Default: 1.0). Float (0.0-100.0). The temperature of the sampling operation. 1 means regular sampling, 0 means always take the highest score, 100.0 is getting closer to uniform probability.}

\item{max_time}{(Default: None). Float (0-120.0). The amount of time in seconds that the query should take maximum. Network can cause some overhead so it will be a soft limit.}

\item{tidy}{Whether to tidy the results into a tibble. Default: TRUE (tidy the results)}

\item{use_gpu}{Whether to use GPU for inference.}

\item{use_cache}{Whether to use cached inference results for previously seen inputs.}

\item{wait_for_model}{Whether to wait for the model to be ready instead of receiving a 503 error after a certain amount of time.}

\item{use_auth_token}{The token to use as HTTP bearer authorization for the Inference API. Defaults to HUGGING_FACE_HUB_TOKEN environment variable.}

\item{stop_on_error}{Whether to throw an error if an API error is encountered. Defaults to FALSE (do not throw error).}

\item{...}{Additional arguments passed internally, including the model object or model ID.}
}
\value{
The results of the inference
}
\description{
Conversational API Inference
}
\seealso{
\url{https://huggingface.co/docs/inference-providers}
}
