% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ez.R
\name{hf_ez_conversational_local_inference}
\alias{hf_ez_conversational_local_inference}
\title{Conversational Local Inference}
\usage{
hf_ez_conversational_local_inference(
  text,
  generated_responses = NULL,
  past_user_inputs = NULL,
  min_length = NULL,
  max_length = NULL,
  top_k = NULL,
  top_p = NULL,
  temperature = 1,
  max_time = NULL,
  tidy = TRUE,
  ...
)
}
\arguments{
\item{text}{The last input from the user in the conversation.}

\item{generated_responses}{A list of strings corresponding to the earlier replies from the model.}

\item{past_user_inputs}{A list of strings corresponding to the earlier replies from the user. Should be of the same length of generated_responses.}

\item{min_length}{(Default: None). Integer to define the minimum length in tokens of the output summary.}

\item{max_length}{(Default: None). Integer to define the maximum length in tokens of the output summary.}

\item{top_k}{(Default: None). Integer to define the top tokens considered within the sample operation to create new text.}

\item{top_p}{(Default: None). Float to define the tokens that are within the sample operation of text generation. Add tokens in the sample for more probable to least probable until the sum of the probabilities is greater than top_p.}

\item{temperature}{(Default: 1.0). Float (0.0-100.0). The temperature of the sampling operation. 1 means regular sampling, 0 means always take the highest score, 100.0 is getting closer to uniform probability.}

\item{max_time}{(Default: None). Float (0-120.0). The amount of time in seconds that the query should take maximum. Network can cause some overhead so it will be a soft limit.}

\item{repetition_penalty}{(Default: None). Float (0.0-100.0). The more a token is used within generation the more it is penalized to not be picked in successive generation passes.}
}
\value{
The results of the inference
}
\description{
Conversational Local Inference
}
\seealso{
\url{https://huggingface.co/docs/transformers/main/en/pipeline_tutorial}
}
