% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ez.R
\name{hf_ez_fill_mask_local_inference}
\alias{hf_ez_fill_mask_local_inference}
\title{Fill Mask Local Inference}
\usage{
hf_ez_fill_mask_local_inference(string, tidy = TRUE, ...)
}
\arguments{
\item{string}{a string to be filled from, must contain the [MASK] token (check model card for exact name of the mask)}

\item{tidy}{Whether to tidy the results into a tibble. Default: TRUE (tidy the results)}
}
\value{
The results of the inference
}
\description{
Fill Mask Local Inference
}
\seealso{
\url{https://huggingface.co/docs/transformers/main/en/pipeline_tutorial}
}
