% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ez.R
\name{hf_ez_question_answering_local_inference}
\alias{hf_ez_question_answering_local_inference}
\title{Question Answering Local Inference}
\usage{
hf_ez_question_answering_local_inference(
  question,
  context,
  flatten = TRUE,
  ...
)
}
\arguments{
\item{question}{a question to be answered based on the provided context}

\item{context}{the context to consult for answering the question}

\item{flatten}{Whether to flatten the results into a data frame. Default: TRUE (flatten the results)}
}
\value{
The results of the inference
}
\description{
Question Answering Local Inference
}
\seealso{
\url{https://huggingface.co/docs/transformers/main/en/pipeline_tutorial}
}
