% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ez.R
\name{hf_ez_sentence_similarity_api_inference}
\alias{hf_ez_sentence_similarity_api_inference}
\title{Sentence Similarity API Inference}
\usage{
hf_ez_sentence_similarity_api_inference(
  source_sentence,
  sentences,
  tidy = TRUE,
  use_gpu = FALSE,
  use_cache = FALSE,
  wait_for_model = FALSE,
  use_auth_token = NULL,
  stop_on_error = FALSE,
  ...
)
}
\arguments{
\item{source_sentence}{The string that you wish to compare the other strings with. This can be a phrase, sentence, or longer passage, depending on the model being used.}

\item{sentences}{A list of strings which will be compared against the source_sentence.}

\item{tidy}{Whether to tidy the results into a tibble. Default: TRUE (tidy the results)}

\item{use_gpu}{Whether to use GPU for inference.}

\item{use_cache}{Whether to use cached inference results for previously seen inputs.}

\item{wait_for_model}{Whether to wait for the model to be ready instead of receiving a 503 error after a certain amount of time.}

\item{use_auth_token}{The token to use as HTTP bearer authorization for the Inference API. Defaults to HUGGING_FACE_HUB_TOKEN environment variable.}

\item{stop_on_error}{Whether to throw an error if an API error is encountered. Defaults to FALSE (do not throw error).}

\item{...}{Additional arguments passed internally, including the model object or model ID.}
}
\value{
The results of the inference
}
\description{
Sentence Similarity API Inference
}
\seealso{
\url{https://huggingface.co/docs/inference-providers}
}
