% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ez.R
\name{hf_ez_text_generation_api_inference}
\alias{hf_ez_text_generation_api_inference}
\title{Text Generation API Inference}
\usage{
hf_ez_text_generation_api_inference(
  string,
  top_k = NULL,
  top_p = NULL,
  temperature = 1,
  repetition_penalty = NULL,
  max_new_tokens = NULL,
  max_time = NULL,
  return_full_text = TRUE,
  num_return_sequences = 1L,
  do_sample = TRUE,
  tidy = TRUE,
  use_gpu = FALSE,
  use_cache = FALSE,
  wait_for_model = FALSE,
  use_auth_token = NULL,
  stop_on_error = FALSE,
  ...
)
}
\arguments{
\item{string}{a string to be generated from}

\item{top_k}{(Default: None). Integer to define the top tokens considered within the sample operation to create new text.}

\item{top_p}{(Default: None). Float to define the tokens that are within the sample operation of text generation. Add tokens in the sample for more probable to least probable until the sum of the probabilities is greater than top_p}

\item{temperature}{Float (0.0-100.0). The temperature of the sampling operation. 1 means regular sampling, 0 means always take the highest score, 100.0 is getting closer to uniform probability. Default: 1.0}

\item{repetition_penalty}{(Default: None). Float (0.0-100.0). The more a token is used within generation the more it is penalized to not be picked in successive generation passes.}

\item{max_new_tokens}{(Default: None). Int (0-250). The amount of new tokens to be generated, this does not include the input length it is a estimate of the size of generated text you want. Each new tokens slows down the request, so look for balance between response times and length of text generated.}

\item{max_time}{(Default: None). Float (0-120.0). The amount of time in seconds that the query should take maximum. Network can cause some overhead so it will be a soft limit. Use that in combination with max_new_tokens for best results.}

\item{return_full_text}{(Default: True). Bool. If set to False, the return results will not contain the original query making it easier for prompting.}

\item{num_return_sequences}{(Default: 1). Integer. The number of proposition you want to be returned.}

\item{do_sample}{(Optional: True). Bool. Whether or not to use sampling, use greedy decoding otherwise.#'}

\item{tidy}{Whether to tidy the results into a tibble. Default: TRUE (tidy the results)}

\item{use_gpu}{Whether to use GPU for inference.}

\item{use_cache}{Whether to use cached inference results for previously seen inputs.}

\item{wait_for_model}{Whether to wait for the model to be ready instead of receiving a 503 error after a certain amount of time.}

\item{use_auth_token}{The token to use as HTTP bearer authorization for the Inference API. Defaults to HUGGING_FACE_HUB_TOKEN environment variable.}

\item{stop_on_error}{Whether to throw an error if an API error is encountered. Defaults to FALSE (do not throw error).}

\item{...}{Additional arguments passed internally, including the model object or model ID.}
}
\value{
The results of the inference
}
\description{
Text Generation API Inference
}
\seealso{
\url{https://huggingface.co/docs/inference-providers}
}
