% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{hf_fill_mask_payload}
\alias{hf_fill_mask_payload}
\title{Fill Mask Task Payload}
\usage{
hf_fill_mask_payload(string)
}
\arguments{
\item{string}{a string to be filled from, must contain the [MASK] token (check model card for exact name of the mask)}
}
\value{
An inference payload
}
\description{
Tries to fill in a hole with a missing word (token to be precise). That’s the base task for BERT models.
}
\examples{
hf_fill_mask_payload("The capital of France is [MASK].")
}
\seealso{
\url{https://huggingface.co/docs/inference-providers/tasks/fill-mask}
}
