% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/huggingface_hub.R
\name{hf_list_authors}
\alias{hf_list_authors}
\title{List Authors}
\usage{
hf_list_authors(pattern = NULL)
}
\arguments{
\item{pattern}{A search term or regular expression. Defaults to NULL (return all results).}
}
\value{
A character vector of matching model authors.
}
\description{
List Model Authors
}
\examples{
\dontrun{
hf_list_authors(pattern = "^sam")
}
}
\seealso{
\url{https://huggingface.co/docs/hub/searching-the-hub}
}
