% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/huggingface_hub.R
\name{hf_list_datasets}
\alias{hf_list_datasets}
\title{List Datasets}
\usage{
hf_list_datasets(pattern = NULL)
}
\arguments{
\item{pattern}{A search term or regular expression. Defaults to NULL (return all results).}
}
\description{
List Model Datasets
}
\examples{
hf_list_datasets("imdb")
}
\seealso{
\url{https://huggingface.co/docs/hub/searching-the-hub}
}
