% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/huggingface_hub.R
\name{hf_list_libraries}
\alias{hf_list_libraries}
\title{List Libraries}
\usage{
hf_list_libraries(pattern = NULL)
}
\arguments{
\item{pattern}{A search term or regular expression. Defaults to NULL (return all results).}
}
\description{
List Model Libraries
}
\examples{
hf_list_libraries('pytorch|tensorflow')
}
\seealso{
\url{https://huggingface.co/docs/hub/searching-the-hub}
}
