% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformers.R
\name{hf_load_AutoModel}
\alias{hf_load_AutoModel}
\title{Import a pre-trained model for a specific task.}
\usage{
hf_load_AutoModel(
  model_type = "AutoModelForSequenceClassification",
  model_id,
  use_auth_token = NULL
)
}
\arguments{
\item{model_type}{The AutoModel's type passed as a string e.g. c("AutoModelForQuestionAnswering", "AutoModelForTokenClassification", "AutoModelForSequenceClassification")}

\item{model_id}{The model's name or id on the Hugging Face hub}

\item{use_auth_token}{For private models, copy and paste your auth token in as a string.}
}
\value{
an AutoMNodel object for a specific task
}
\description{
Function differs from `hf_load_model` in that `hf_load_model` currently only loads AutoModels i.e. not AutoModelsForX
}
\seealso{
\url{https://huggingface.co/transformers/v3.0.2/model_doc/auto.html}
}
