% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformers.R
\name{hf_load_AutoModel_for_task}
\alias{hf_load_AutoModel_for_task}
\title{Import a pre-trained model for a specific task.}
\usage{
hf_load_AutoModel_for_task(
  model_type = "AutoModelForSequenceClassification",
  model_id,
  use_auth_token = NULL
)
}
\arguments{
\item{model_type}{The AutoModel's type passed as a string e.g. c("AutoModelForQuestionAnswering", "AutoModelForTokenClassification", "AutoModelForSequenceClassification")}

\item{model_id}{The model's name or id on the Hugging Face hub}

\item{use_auth_token}{For private models, copy and paste your auth token in as a string.}
}
\value{
an AutoModel object for a specific task
}
\description{
Use this function when you need to load an AutoModel for a specific task separate to a pipeline.
The model's config should come ready-equipped for a specific task according to what you input as model_type.
}
\seealso{
\url{https://huggingface.co/transformers/v3.0.2/model_doc/auto.html}
}
