% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{hf_load_dataset}
\alias{hf_load_dataset}
\title{Load dataset
Load dataset from Huggingface}
\usage{
hf_load_dataset(dataset, split = NULL, as_tibble = FALSE, ...)
}
\arguments{
\item{dataset}{The name of a huggingface dataset. Use hf_list_models() to find a dataset.}
}
\value{
A Huggingface dataset.
}
\description{
Load dataset
Load dataset from Huggingface
}
\examples{
# Retrieve the 'emotion' dataset
emotion <- hf_load_dataset("emotion")
emotion
# Extract and visualize the training split in the emotion data
hf_load_dataset("emotion", as_tibble = TRUE, split = "train") \%>\%
  add_count(label) \%>\%
  mutate(
    label = fct_reorder(as.factor(label), n)
  ) \%>\%
  ggplot(aes(label)) +
  geom_bar()
}
\seealso{
\url{https://huggingface.co/docs/datasets/index}
}
