% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{hf_python_depends}
\alias{hf_python_depends}
\title{Install Python Dependencies}
\usage{
hf_python_depends(
  packages = c("transformers", "sentencepiece", "huggingface_hub", "datasets",
    "sentence-transformers")
)
}
\arguments{
\item{packages}{Python libraries needed for local model usage. \cr
Defaults to transformers, sentencepiece, huggingface_hub, datasets, and sentence-transformers.}
}
\description{
Installs python packages needed to run huggingfaceR functions
}
