% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentence-transformers.R
\name{hf_sentence_encode}
\alias{hf_sentence_encode}
\title{Use a Sentence Transformers pipeline to extract document(s)/sentence(s) embedding(s)}
\usage{
hf_sentence_encode(
  model,
  text,
  batch_size = 64L,
  show_progress_bar = TRUE,
  tidy = TRUE,
  ...
)
}
\arguments{
\item{model}{Model object you loaded with `hf_load_sentence_model()`}

\item{text}{The text, or texts, you wish to embed/encode.}

\item{batch_size}{How many texts to embed at once.}

\item{show_progress_bar}{Whether to print a progress bar in the console or not.}

\item{tidy}{Whether to tidy the output into a tibble or not.}

\item{...}{other args sent to the model's encode method, e.g. device = device}
}
\value{
An embedding
}
\description{
Use a Sentence Transformers pipeline to extract document(s)/sentence(s) embedding(s)
}
\examples{
\dontrun{
text <- c("There are things we do know, things we don't know, and then there is quantum mechanics.")
sentence_mod <- hf_load_sentence_model("paraphrase-MiniLM-L6-v2")
embeddings <- hf_sentence_encode(model = sentence_mod, text, show_progress_bar = TRUE)
}
}
