% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{hf_table_question_answering_payload}
\alias{hf_table_question_answering_payload}
\title{Table Question Answering Payload}
\usage{
hf_table_question_answering_payload(query, table)
}
\arguments{
\item{query}{The query in plain text that you want to ask the table}

\item{table}{A table of data represented as a dict of list where entries are headers and the lists are all the values, all lists must have the same size.}
}
\value{
An inference payload
}
\description{
Don’t know SQL? Don’t want to dive into a large spreadsheet? Ask questions in plain english!
}
\examples{
hf_table_question_answering_payload(
  query = "How many rows are shown?",
  table = list(name = c("Alice", "Bob"), rows = c("1", "2"))
)
}
\seealso{
\url{https://huggingface.co/docs/inference-providers/tasks/table-question-answering}
}
