% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{hf_translation_payload}
\alias{hf_translation_payload}
\title{Translation Payload}
\usage{
hf_translation_payload(string)
}
\arguments{
\item{string}{a string to be translated in the original languages}
}
\value{
An inference payload
}
\description{
This task is well known to translate text from one language to another
}
\examples{
hf_translation_payload("Hello, world.")
}
\seealso{
\url{https://huggingface.co/docs/inference-providers/tasks/translation}
}
