% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_with_downloads.R
\docType{data}
\name{models_with_downloads}
\alias{models_with_downloads}
\title{Dataset of model names, tasks & download info}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 55524 rows and 5 columns.
}
\usage{
models_with_downloads
}
\description{
\describe{
#' \item{model}{name of model}
#' \item{downloads}{number of downloads}
#' \item{task}{task model performsl}
#' \item{sha}{model's secure hash algorithm}
#' \item{private}{whether model is public or private}
}
}
\keyword{datasets}
