% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{null_coalesce}
\alias{null_coalesce}
\alias{\%||\%}
\title{Null-coalescing operator}
\usage{
x \%||\% y
}
\arguments{
\item{x}{First value}

\item{y}{Second value (default)}
}
\description{
Null-coalescing operator
}
\keyword{internal}
