% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_reticulate_for_transformers.R
\name{setup_reticulate_for_transformers}
\alias{setup_reticulate_for_transformers}
\title{Setup miniconda to work with the hugging face transformer python library}
\usage{
setup_reticulate_for_transformers()
}
\description{
Setup miniconda to work with the hugging face transformer python library
}
\examples{
setup_reticulate_for_transformers()
}
