% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ez.R
\name{hf_ez_conversational}
\alias{hf_ez_conversational}
\title{Create a Conversational Agent}
\usage{
hf_ez_conversational(model_id = "microsoft/DialoGPT-large", use_api = FALSE)
}
\arguments{
\item{model_id}{A model_id. Run hf_search_models(...) for model_ids. Defaults to 'microsoft/DialoGPT-large'.}

\item{use_api}{Whether to use the Inference API to run the model (TRUE) or download and run the model locally (FALSE). Defaults to FALSE}
}
\value{
A conversational object
}
\description{
This task corresponds to any chatbot like structure. Models tend to have shorter max_length, so please check with caution when using a given model if you need long range dependency or not.
}
\examples{
\dontrun{
# Load the default translation model
chat_bot <- hf_ez_conversational()

# Translate from Engllish to Spanish.
chat_bot$infer(string = "Hi, I recently bought a device from your company but it is not working as advertised and I would like to get reimbursed!", candidate_labels = c("refund", "legal", "faq"))
}
}
\seealso{
\url{https://huggingface.co/docs/api-inference/detailed_parameters#zero-shot-classification-task}
}
