% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ez.R
\name{hf_ez_fill_mask}
\alias{hf_ez_fill_mask}
\title{Perform Fill-in-the-Blank Tasks}
\usage{
hf_ez_fill_mask(model_id = "google-bert/bert-base-uncased", use_api = FALSE)
}
\arguments{
\item{model_id}{A model_id. Run hf_search_models(...) for model_ids. Defaults to 'bert-base-uncased'.}

\item{use_api}{Whether to use the Inference API to run the model (TRUE) or download and run the model locally (FALSE). Defaults to FALSE}
}
\value{
A fill mask object
}
\description{
Tries to fill in a hole with a missing word (token to be precise). That’s the base task for BERT models.
}
\examples{
\dontrun{
# Load the default model and use local inference
ez <- hf_ez_fill_mask()
ez$infer(string = "The answer to the universe is [MASK].")

# Load a specific model and use the api for inference. Note the mask is different for different models.
ez <- hf_ez_fill_mask(model_id = 'xlm-roberta-base', use_api = TRUE)
ez$infer(string = "The answer to the universe is <MASK>.")
}
}
\seealso{
\url{https://huggingface.co/docs/api-inference/detailed_parameters#fill-mask-task}
}
