% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ez.R
\name{hf_ez_question_answering}
\alias{hf_ez_question_answering}
\title{EZ Question Answering}
\usage{
hf_ez_question_answering(
  model_id = "deepset/roberta-base-squad2",
  use_api = FALSE
)
}
\arguments{
\item{model_id}{A model_id. Run hf_search_models(...) for model_ids. Defaults to 'deepset/roberta-base-squad2'.}

\item{use_api}{Whether to use the Inference API to run the model (TRUE) or download and run the model locally (FALSE). Defaults to FALSE}
}
\value{
A question answering object
}
\description{
Want to have a nice know-it-all bot that can answer any question?
}
\seealso{
\url{https://huggingface.co/docs/api-inference/detailed_parameters#question-answering-task}
}
