% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ez.R
\name{hf_ez_sentence_similarity}
\alias{hf_ez_sentence_similarity}
\title{EZ Sentence Similarity}
\usage{
hf_ez_sentence_similarity(
  model_id = "sentence-transformers/all-MiniLM-L6-v2",
  use_api = FALSE
)
}
\arguments{
\item{model_id}{A model_id. Run hf_search_models(...) for model_ids. Defaults to 'sentence-transformers/all-MiniLM-L6-v2'.}

\item{use_api}{Whether to use the Inference API to run the model (TRUE) or download and run the model locally (FALSE). Defaults to FALSE}
}
\value{
A sentence similarity object
}
\description{
Calculate the semantic similarity between one text and a list of other sentences by comparing their embeddings.
}
\seealso{
\url{https://huggingface.co/docs/api-inference/detailed_parameters#sentence-similarity-task}
}
