% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ez.R
\name{hf_ez_sentence_similarity}
\alias{hf_ez_sentence_similarity}
\title{Compare Sentence Similarity Semantically}
\usage{
hf_ez_sentence_similarity(
  model_id = "sentence-transformers/all-MiniLM-L6-v2",
  use_api = FALSE
)
}
\arguments{
\item{model_id}{A model_id. Run hf_search_models(...) for model_ids. Defaults to 'sentence-transformers/all-MiniLM-L6-v2'.}

\item{use_api}{Whether to use the Inference API to run the model (TRUE) or download and run the model locally (FALSE). Defaults to FALSE}
}
\value{
A sentence similarity object
}
\description{
Calculate the semantic similarity between one text and a list of other sentences by comparing their embeddings.
}
\examples{
\dontrun{
# Load the default model and use local inference
ez <- hf_ez_sentence_similarity()
ez$infer(source_sentence = 'That is a happy person', sentences = list('That is a happy dog', 'That is a very happy person", "Today is a sunny day'))

# Use the api for inference.
ez <- hf_ez_sentence_similarity(use_api = TRUE)
ez$infer(source_sentence = 'That is a happy person', sentences = list('That is a happy dog', 'That is a very happy person', 'Today is a sunny day'))
}
}
\seealso{
\url{https://huggingface.co/docs/api-inference/detailed_parameters#sentence-similarity-task}
}
