% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ez.R
\name{hf_ez_summarization}
\alias{hf_ez_summarization}
\title{EZ Summarization}
\usage{
hf_ez_summarization(model_id = "facebook/bart-large-cnn", use_api = FALSE)
}
\arguments{
\item{model_id}{A model_id. Run hf_search_models(...) for model_ids. Defaults to 'facebook/bart-large-cnn'.}

\item{use_api}{Whether to use the Inference API to run the model (TRUE) or download and run the model locally (FALSE). Defaults to FALSE}
}
\value{
A summarization object
}
\description{
This task is well known to summarize longer text into shorter text. Be careful, some models have a maximum length of input. That means that the summary cannot handle full books for instance. Be careful when choosing your model.
}
\seealso{
\url{https://huggingface.co/docs/api-inference/detailed_parameters#summarization-task}
}
