% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ez.R
\name{hf_ez_table_question_answering}
\alias{hf_ez_table_question_answering}
\title{Answer Questions about a Data Table}
\usage{
hf_ez_table_question_answering(
  model_id = "google/tapas-base-finetuned-wtq",
  use_api = FALSE
)
}
\arguments{
\item{model_id}{A model_id. Run hf_search_models(...) for model_ids. Defaults to 'google/tapas-base-finetuned-wtq'.}

\item{use_api}{Whether to use the Inference API to run the model (TRUE) or download and run the model locally (FALSE). Defaults to FALSE}
}
\value{
A table question answering object
}
\description{
Don’t know SQL? Don’t want to dive into a large spreadsheet? Ask questions in plain english!
}
\seealso{
\url{https://huggingface.co/docs/api-inference/detailed_parameters#table-question-answering-task}
}
