% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ez.R
\name{hf_ez_table_question_answering_local_inference}
\alias{hf_ez_table_question_answering_local_inference}
\title{Table Question Answering Local Inference}
\usage{
hf_ez_table_question_answering_local_inference(query, table, tidy = TRUE, ...)
}
\arguments{
\item{query}{The query in plain text that you want to ask the table}

\item{table}{A dataframe with all text columns.}

\item{tidy}{Whether to tidy the results into a tibble. Default: TRUE (tidy the results)}
}
\value{
The results of the inference
}
\description{
Table Question Answering Local Inference
}
\seealso{
\url{https://huggingface.co/docs/transformers/main/en/pipeline_tutorial}
}
