% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ez.R
\name{hf_ez_text2text_generation}
\alias{hf_ez_text2text_generation}
\title{Answer General Questions about a Text}
\usage{
hf_ez_text2text_generation(model_id = "google/flan-t5-small", use_api = FALSE)
}
\arguments{
\item{model_id}{A model_id. Run hf_search_models(...) for model_ids. Defaults to 'google/flan-t5-small'.}

\item{use_api}{Whether to use the Inference API to run the model (TRUE) or download and run the model locally (FALSE). Defaults to FALSE}
}
\value{
A text2text generation object
}
\description{
Essentially Text-generation task. But uses Encoder-Decoder architecture, so might change in the future for more options.
}
\seealso{
\url{https://huggingface.co/docs/api-inference/detailed_parameters#text2text-generation-task}
}
