% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ez.R
\name{hf_ez_text_classification_local_inference}
\alias{hf_ez_text_classification_local_inference}
\title{Text Classification Local Inference}
\usage{
hf_ez_text_classification_local_inference(string, tidy = TRUE, ...)
}
\arguments{
\item{string}{a string to be classified}

\item{tidy}{Whether to tidy the results into a tibble. Default: TRUE (tidy the results)}
}
\value{
The results of the inference
}
\description{
Text Classification Local Inference
}
\seealso{
\url{https://huggingface.co/docs/transformers/main/en/pipeline_tutorial}
}
