% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ez.R
\name{hf_ez_token_classification_api_inference}
\alias{hf_ez_token_classification_api_inference}
\title{Token Classification API Inference}
\usage{
hf_ez_token_classification_api_inference(
  string,
  tidy = TRUE,
  use_gpu = FALSE,
  use_cache = FALSE,
  wait_for_model = FALSE,
  use_auth_token = NULL,
  stop_on_error = FALSE,
  ...
)

hf_ez_token_classification_api_inference(
  string,
  tidy = TRUE,
  use_gpu = FALSE,
  use_cache = FALSE,
  wait_for_model = FALSE,
  use_auth_token = NULL,
  stop_on_error = FALSE,
  ...
)
}
\arguments{
\item{string}{a string to be translated}

\item{tidy}{Whether to tidy the results into a tibble. Default: TRUE (tidy the results)}

\item{use_gpu}{Whether to use GPU for inference.}

\item{use_cache}{Whether to use cached inference results for previously seen inputs.}

\item{wait_for_model}{Whether to wait for the model to be ready instead of receiving a 503 error after a certain amount of time.}

\item{use_auth_token}{The token to use as HTTP bearer authorization for the Inference API. Defaults to HUGGING_FACE_HUB_TOKEN environment variable.}

\item{stop_on_error}{Whether to throw an error if an API error is encountered. Defaults to FALSE (do not throw error).}

\item{aggregation_strategy}{(Default: simple). There are several aggregation strategies. \cr
none: Every token gets classified without further aggregation.  \cr
simple: Entities are grouped according to the default schema (B-, I- tags get merged when the tag is similar).  \cr
first: Same as the simple strategy except words cannot end up with different tags. Words will use the tag of the first token when there is ambiguity.  \cr
average: Same as the simple strategy except words cannot end up with different tags. Scores are averaged across tokens and then the maximum label is applied.  \cr
max: Same as the simple strategy except words cannot end up with different tags. Word entity will be the token with the maximum score.  \cr}
}
\value{
The results of the inference

The results of the inference
}
\description{
Token Classification API Inference

Translation API Inference
}
\seealso{
\url{https://huggingface.co/docs/api-inference/index}

\url{https://huggingface.co/docs/api-inference/index}
}
