% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ez.R
\name{hf_ez_translation}
\alias{hf_ez_translation}
\title{Translate between Languages}
\usage{
hf_ez_translation(model_id = "Helsinki-NLP/opus-mt-en-es", use_api = FALSE)
}
\arguments{
\item{model_id}{A model_id. Run hf_search_models(...) for model_ids. Defaults to 'Helsinki-NLP/opus-mt-en-es'.}

\item{use_api}{Whether to use the Inference API to run the model (TRUE) or download and run the model locally (FALSE). Defaults to FALSE}
}
\value{
A translation object
}
\description{
This task is well known to translate text from one language to another
}
\examples{
\dontrun{
# Load the default translation model
ez_model <- hf_ez_translation()

# Translate from English to Spanish.
ez_model$infer(string = "My name is Sarah and I live in London")
}
}
\seealso{
\url{https://huggingface.co/docs/api-inference/detailed_parameters#translation-task}
}
