% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ez.R
\name{hf_ez_zero_shot_classification}
\alias{hf_ez_zero_shot_classification}
\title{Perform Text Classification with No Context Required}
\usage{
hf_ez_zero_shot_classification(
  model_id = "facebook/bart-large-mnli",
  use_api = FALSE
)
}
\arguments{
\item{model_id}{A model_id. Run hf_search_models(...) for model_ids. Defaults to 'facebook/bart-large-mnli'.}

\item{use_api}{Whether to use the Inference API to run the model (TRUE) or download and run the model locally (FALSE). Defaults to FALSE}
}
\value{
A zero shot classification object
}
\description{
This task is super useful to try out classification with zero code, you simply pass a sentence/paragraph and the possible labels for that sentence, and you get a result.
}
\examples{
\dontrun{
# Load the default translation model
zs <- hf_ez_zero_shot_classification()

# Translate from Engllish to Spanish.
zs$infer(string = "Hi, I recently bought a device from your company but it is not working as advertised and I would like to get reimbursed!", candidate_labels = c("refund", "legal", "faq"))
}
}
\seealso{
\url{https://huggingface.co/docs/api-inference/detailed_parameters#zero-shot-classification-task}
}
