% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searching_huggingface.R
\name{hf_list_authors}
\alias{hf_list_authors}
\title{List Authors}
\usage{
hf_list_authors(pattern = NULL)
}
\arguments{
\item{pattern}{A search term or regular expression. Defaults to NULL (return all results).}
}
\description{
List Model Authors
}
\examples{
hf_list_authors(pattern = '^sam')
}
\seealso{
\url{https://huggingface.co/docs/hub/searching-the-hub}
}
