% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/huggingface_hub.R
\name{hf_list_licenses}
\alias{hf_list_licenses}
\title{List Licenses}
\usage{
hf_list_licenses(pattern = NULL)
}
\arguments{
\item{pattern}{A search term or regular expression. Defaults to NULL (return all results).}
}
\description{
List Model Licenses
}
\examples{
hf_list_licenses('mit')
}
\seealso{
\url{https://huggingface.co/docs/hub/searching-the-hub}
}
