% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searching_huggingface.R
\name{hf_list_tasks}
\alias{hf_list_tasks}
\title{List Tasks}
\usage{
hf_list_tasks(pattern = NULL)
}
\arguments{
\item{pattern}{A search term or regular expression. Defaults to NULL (return all results).}
}
\description{
List Model Tasks
}
\examples{
hf_list_tasks('^image.*tion')
}
