% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{hf_load_dataset}
\alias{hf_load_dataset}
\title{Load dataset
Load dataset from Huggingface}
\usage{
hf_load_dataset(dataset, split = NULL, as_tibble = FALSE, ...)
}
\arguments{
\item{dataset}{The name of a huggingface dataset. Use hf_list_models() to find a dataset.}

\item{split}{can be either 'train', 'test', 'validation', or left NULL for all three.}

\item{as_tibble}{defaults to FALSE. Set to TRUE to return a tibble.}

\item{...}{fed to load_dataset()}
}
\value{
A Huggingface dataset as a tibble or as it's default arrow dataset.
}
\description{
Load dataset
Load dataset from Huggingface
}
\seealso{
\url{https://huggingface.co/docs/datasets/index}
}
