% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{hf_load_dataset}
\alias{hf_load_dataset}
\title{Load a dataset from the Hugging Face Hub!}
\usage{
hf_load_dataset(
  dataset,
  split = NULL,
  label_conversion = c("str2int", "int2str")
)
}
\arguments{
\item{dataset}{The name of a Hugging Face dataset saved on the Hub. Use hf_list_models() to find a dataset.}

\item{split}{Usually one of 'train' , 'test', 'validation' however, check the dataset's meta data at The Hub first.}

\item{label_conversion}{Whether to add an additional column converting labels from str2int or int2str?}
}
\value{
A Hugging Face data set as a tibble
}
\description{
Function has multiple uses - getting pre-made datasets for exploratory analysis, or to figure as means for evaluating your fine-tuned models.
}
\examples{
(emo_all_splits <- hf_load_dataset('emo'))

(imdb_train <- hf_load_dataset('imdb', split = "train"))


}
\seealso{
\url{https://huggingface.co/docs/datasets/index}
}
