% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformers.R
\name{hf_load_model}
\alias{hf_load_model}
\title{Load Model
Load Model from Huggingface}
\usage{
hf_load_model(model_id, tokenizer = NULL, task = NULL)
}
\arguments{
\item{model_id}{The id of the model given in the url by https://huggingface.co/model_name.}

\item{tokenizer}{The tokenizer function used to tokenize inputs. Defaults to NULL (one will be automatically loaded).}

\item{task}{The task the model will accomplish. Run hf_list_tasks() for options.}
}
\value{
A Huggingface model ready for prediction.
}
\description{
Load Model
Load Model from Huggingface
}
\examples{
model <- hf_load_model('facebook/bart-large-mnli')
model$task
model("Joe is eating a donut and enjoying himself.", c("happy", "neutral", "sad"))
}
\seealso{
\url{https://huggingface.co/docs/transformers/main/en/pipeline_tutorial}
}
