% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformers.R
\name{hf_load_model}
\alias{hf_load_model}
\title{Load a pre-trained AutoModel object from Hugging Face}
\usage{
hf_load_model(model_id, ...)
}
\arguments{
\item{model_id}{model_id The id of the model given in the url by https://huggingface.co/model_name.}

\item{...}{sent to AutoModel.from_pretrained()}
}
\value{
a pre-trained model object
}
\description{
Load a pre-trained AutoModel object from Hugging Face
}
\examples{
\dontrun{
model <- hf_load_model("distilbert-base-uncased")
}
}
