% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentence-transformers.R
\name{hf_load_sentence_model}
\alias{hf_load_sentence_model}
\title{Load Sentence Model}
\usage{
hf_load_sentence_model(model_id)
}
\arguments{
\item{model_id}{The id of a sentence-transformers model. Use hf_search_models(author = 'sentence-transformers') to find suitable models.}
}
\value{
A Huggingface model ready for prediction.
}
\description{
Load Sentence Transformers Model from Huggingface
}
\seealso{
\url{https://huggingface.co/sentence-transformers}
}
