% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{hf_question_answering_payload}
\alias{hf_question_answering_payload}
\title{Question Answering Payload}
\usage{
hf_question_answering_payload(question, context)
}
\arguments{
\item{question}{a question to be answered based on the provided context}

\item{context}{the context to consult for answering the question}
}
\value{
An inference payload
}
\description{
Want to have a nice know-it-all bot that can answer any question?
}
\seealso{
\url{https://huggingface.co/docs/api-inference/detailed_parameters#question-answering-task}
}
