% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/huggingface_hub.R
\name{hf_search_models}
\alias{hf_search_models}
\title{Search Models}
\usage{
hf_search_models(
  author = NULL,
  language = NULL,
  library = NULL,
  name = NULL,
  tags = NULL,
  task = NULL,
  dataset = NULL
)
}
\arguments{
\item{author}{Filter by model author. Run hf_list_authors() for options.}

\item{language}{Filter by the languages the model accommodates. Run hf_list_languages() for options.}

\item{library}{Filter by the deep learning libraries which work with the model. Run hf_list_libraries() for options.}

\item{name}{Filter by model names. Run hf_list_models() for options.}

\item{tags}{Filter by model tags.}

\item{task}{Filter by tasks the model can accomplish. Run hf_list_tasks() for options.}

\item{dataset}{Filter by the datasets the model was trained on. hf_list_datasets()}
}
\description{
Search Huggingface Models
}
\examples{
hf_search_models(library = "pytorch", dataset = "mnli")
hf_search_models(author = "facebook", name = "bart")
}
\seealso{
\url{https://huggingface.co/docs/hub/searching-the-hub}
}
