% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{hf_text_classification_payload}
\alias{hf_text_classification_payload}
\title{Text Classification Payload}
\usage{
hf_text_classification_payload(string)
}
\arguments{
\item{string}{a string to be classified}
}
\value{
An inference payload
}
\description{
Usually used for sentiment-analysis this will output the likelihood of classes of an input.
}
\seealso{
\url{https://huggingface.co/docs/api-inference/detailed_parameters#sentence-similarity-task}
}
