% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{hf_token_classification_payload}
\alias{hf_token_classification_payload}
\title{Token Classification Payload}
\usage{
hf_token_classification_payload(string, aggregation_strategy = "simple")
}
\arguments{
\item{string}{a string to be classified}

\item{aggregation_strategy}{(Default: simple). There are several aggregation strategies. \cr
none: Every token gets classified without further aggregation.  \cr
simple: Entities are grouped according to the default schema (B-, I- tags get merged when the tag is similar).  \cr
first: Same as the simple strategy except words cannot end up with different tags. Words will use the tag of the first token when there is ambiguity.  \cr
average: Same as the simple strategy except words cannot end up with different tags. Scores are averaged across tokens and then the maximum label is applied.  \cr
max: Same as the simple strategy except words cannot end up with different tags. Word entity will be the token with the maximum score.  \cr}
}
\value{
An inference payload
}
\description{
Usually used for sentence parsing, either grammatical, or Named Entity Recognition (NER) to understand keywords contained within text.
}
\seealso{
\url{https://huggingface.co/docs/api-inference/detailed_parameters#token-classification-task}
}
