% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{hf_zero_shot_classification_payload}
\alias{hf_zero_shot_classification_payload}
\title{Zero Shot Classification Payload}
\usage{
hf_zero_shot_classification_payload(
  string,
  candidate_labels,
  multi_label = FALSE
)
}
\arguments{
\item{string}{a string to be translated in the original languages}

\item{candidate_labels}{a list of strings that are potential classes for inputs. (max 10 candidate_labels, for more, simply run multiple requests, results are going to be misleading if using too many candidate_labels anyway. If you want to keep the exact same, you can simply run multi_label=True and do the scaling on your end. )}

\item{multi_label}{(Default: false) Boolean that is set to True if classes can overlap}
}
\value{
An inference payload
}
\description{
This task is super useful to try out classification with zero code, you simply pass a sentence/paragraph and the possible labels for that sentence, and you get a result.
}
\seealso{
\url{https://huggingface.co/docs/api-inference/detailed_parameters#zeroshot-classification-task}
}
